Lua on ESP32: the runtime, the pet.* binding, and the demo creature
Brings in the Lua port produced by a parallel task, verified here
against the panel-profile display driver rather than in isolation.

The demo creature script (ADR 0018) now loads and runs on the
esp32s3 target, driving the same live kf_pet_state the ADR 0025 pet
session already ticks. Lua gets a real ESP-IDF component at
ports/esp32/components/lua, reusing cmake/fetch_lua.cmake's own
FetchContent call and source list via include() rather than a
hand-copied file list, so the desktop and device builds cannot drift
apart on which Lua they are compiling.

Two toolchain problems were found and fixed on the way, both worth
knowing about because neither is guessable:

- A plain CMake target glued into an ESP-IDF component via
  target_link_libraries(INTERFACE) lands at the very end of the link
  line, after libcxx.a's single left-to-right archive scan has already
  finished, producing "undefined reference to __wrap_longjmp". Fixed by
  having the component compile its own sources into a real archive.
- xtensa-esp-elf's <limits.h> chain defines LLONG_MAX for C only, never
  for C++. Fixed with LUA_32BITS, applied identically to the component
  and to main -- they must match, since LUA_INTEGER crosses the real API
  boundary in lua_pushinteger/luaL_checkinteger.

Also fixes three genuine -Werror=format bugs in kf_lua_port.cpp and
kf_lua_alloc.cpp: %u paired with uint32_t, which is long unsigned int
on this toolchain but not on desktop. No behaviour change on desktop.

Cost: +155KB of flash, all .text/.rodata executed in place. Internal
SRAM grows by 192 bytes -- two file-scope structs. Free heap and PSRAM
are unchanged, because kf_arena_init_all() has carved out the 1MB
KF_ARENA_LUA block on every ESP32 build since ADR 0020 whether or not
Lua existed; this slice only puts something real inside a reservation
that was already being paid for. kf/budget.h is untouched.

Verified here: idf.py build clean for esp32s3 against ESP-IDF v6.0.2,
zero warnings, firmware 0x6d0b0 bytes with 57% of the app partition
free. Desktop headless builds clean on macOS/clang and all 13 ctest
cases pass. NOT verified on hardware -- whether kf.log() lines reach
the serial console and whether pet.* reads the live session on-device
are open until the next flash.

+++ simulator/src/lua/kf_lua_alloc.cpp
@@ -231,7 +231,7 @@ void kf_lua_alloc_init() {
 
     KF_LOGI(TAG, "ready: %u bytes usable for Lua's heap (KF_ARENA_LUA_BYTES "
                  "%u, %zu bytes of that is this allocator's own bookkeeping)",
-            payload, static_cast<unsigned>(KF_ARENA_LUA_BYTES),
+            static_cast<unsigned>(payload), static_cast<unsigned>(KF_ARENA_LUA_BYTES),
             2u * kTagBytes);
 }
 
@@ -317,7 +317,7 @@ void *kf_lua_alloc(void *ud, void *ptr, size_t osize, size_t nsize) {
                 "out of Lua heap on grow: wanted %zu bytes, had %u. "
                 "Returning NULL, which Lua surfaces as a catchable "
                 "script-level error.",
-                nsize, header->payload_bytes);
+                nsize, static_cast<unsigned>(header->payload_bytes));
         return nullptr;
     }
     const size_t copy_bytes =
